requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
Django==5.0.6
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-documentai==2.29.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
html5lib==1.1
idna==3.7
lxml==5.3.0
mysqlclient==2.2.4
oscrypto==1.3.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1